Tidy! chown + helpers
COPY + chmod was used, instead of the COPY's --chown parameter. This is
a best practice and optimization.

Also, to help with some downstream images (and a common pattern), the
helper:

download_verify_install script has been added. It takes 3 parameters,
the url, hash, and name inside the archive that will be moved into the
/usr/local/bin directory.