Locally building images matches upstream names
For the directory tree:
dev/
dev/000-base
dev/python
dev/node

Previously, the local images would be of the form `dev-000-base`
`dev-python`, ... etc and the upstream images would match the directory
names. HOWEVER the github actions would have to use dev-python as their
targets -- this was weird.

This fixes that, so everything that isn't the composite dockerfile
(multistage builds don't support build targets with slashes) or the
builder of the multistage builds doesn't need to be aware of the
hyphenated target names.

This cleans up the actions, and the locally built images as the "dashes
for slashes" pattern was unnecessarily wacky.